feat(project): Add react-navigation
