+++ package.json
@@ -19,6 +19,7 @@
     "jest-expo": "27.0.0",
     "prettier-eslint": "8.8.2",
     "react-native-scripts": "1.14.0",
+    "react-navigation": "^2.5.5",
     "react-test-renderer": "16.3.1"
   },
   "main": "./node_modules/react-native-scripts/build/bin/crna-entry.js",
