+++ screens/Form.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import { Text, View } from 'react-native'
+
+const Form = () => (
+  <View>
+    <Text>Contação</Text>
+  </View>
+)
+
+export default Form
